move spec as well
